requirements.txt
pyTelegramBotAPI == 4.5.1
requests
python-telegram-bot-calendar
peewee == 3.14.10